+++ frontend/src/App.js
@@ -1,9 +1,9 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import CourseDescriptionIndexPage from "main/pages/CourseDescriptions/CourseDescriptionIndexPage";
 import ProfilePage from "main/pages/ProfilePage";
-import AdminUsersPage from "main/pages/AdminUsersPage";
-import AdminLoadSubjectsPage from "main/pages/AdminLoadSubjectsPage";
-import AdminJobsPage from "main/pages/AdminJobsPage";
+import AdminUsersPage from "main/pages/Admin/AdminUsersPage";
+import AdminLoadSubjectsPage from "main/pages/Admin/AdminLoadSubjectsPage";
+import AdminJobsPage from "main/pages/Admin/AdminJobsPage";
 import DeveloperPage from "main/pages/DeveloperPage"; // route from /developer to DeveloperPage
 
 import { hasRole, useCurrentUser } from "main/utils/currentUser";
