+++ frontend/src/tests/pages/AdminUsersPage.test.js
@@ -8,7 +8,7 @@ import mockConsole from "jest-mock-console";
 import { apiCurrentUserFixtures } from "fixtures/currentUserFixtures";
 import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
 import usersFixtures from "fixtures/usersFixtures";
-import AdminUsersPage from "main/pages/AdminUsersPage";
+import AdminUsersPage from "main/pages/Admin/AdminUsersPage";
 
 describe("AdminUsersPage tests", () => {
   const axiosMock = new AxiosMockAdapter(axios);
